requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
psycopg2-binary>=2.9.0
prefect>=2.14.0
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=8.0.0
